refactor: move registering of template to component module
